Clean up code and fix donebutton styling
Many 'if' statements were introduced that checked "typeof"
unnecessarily. Also, 'introjs-donebutton' was no longer being added to
the "skip" button when we were in the last step, so there was no way to
differentiate it between the "back" button. Both of these problems were
introduced by PikachuEXE.

There was code introduced by afshinm that I cleaned up by making an
intermediate variable so that I didn't have to refer to
this._introItems[this._currentStep] in three locations.