Add option to individual steps: skipButtonLabel
When specified, this will always be the label for the skip button for
that particular step. This was useful for me in a specific case where
IntroJS points to an element that the user has to click in order to be
considered "done" with a section of the tutorial. Because I divide
tutorials into sections, IntroJS doesn't know that I'm only done with
the current set of steps, not ALL of the steps, so it didn't treat it as
a "skip" action. I treat the presence of skipButtonLabel as an indicator
that we should really skip even if we're on the last step.