Refactor skipButtonLabel
It used to do two things: rename skipLabel and also always treat the
"skip" button as "skip" even when on the last step. However, I no longer
need the ability to rename the button, so I turned it into
preventTreatingSkipAsDone, that way it's obvious that the skip button is
JUST a skip button.