Add options for bypassing caching
I ran into a problem due to using IntroJS with React components where
IntroJS caches elements as soon as the steps start, but React may
re-render such elements, causing the DOM references to be different from
what IntroJS has. In these cases, we want to avoid caching so that we
can still point to the correct elements on the screen.

I added a global option, bypassCachingByDefault, and a per-step option,
bypassCaching.