Fix bug around resizing window
This is a very specific bug where you have an element that shows in a
scrollable panel, but only when a window-size-based media query is true.
In that case, IntroJS wouldn't try scrolling the element's parent to put
the element into view.

I don't think the need for this functionality will be common, so it's an
option on the steps themselves called scrollParentOnBrowserResize.